+++ src/main/kotlin/xyz/cssxsh/bilibili/data/dynamic/DynamicText.kt
@@ -16,6 +16,7 @@ data class DynamicText(
     @SerialName("extension")
     private val extension: JsonObject? = null
 ) {
+
     @Serializable
     data class Item(
         @SerialName("at_uids")
@@ -33,7 +34,7 @@ data class DynamicText(
         @SerialName("rp_id")
         private val replyId: Long,
         @SerialName("timestamp")
-        val timestamp: Int,
+        val timestamp: Long,
         @SerialName("uid")
         val uid: Long
     )
